+++ src/main/java/it/univaq/esc/controller/effettuaPrenotazione/EffettuaPrenotazioneImpiantoState.java
@@ -255,7 +255,7 @@ public class EffettuaPrenotazioneImpiantoState extends EffettuaPrenotazioneState
 	 * esistente associato ad una prenotazione di tipo IMPIANTO.
 	 */
 	@Override
-	public Object aggiungiPartecipanteAEventoEsistente(Long idEvento, String emailPartecipante) {
+	public Object aggiungiPartecipanteAEventoEsistente(Integer idEvento, String emailPartecipante) {
 		Appuntamento appuntamento = this.getRegistroAppuntamenti().getAppuntamentoById(idEvento);
 		if (appuntamento != null) {
 			this.aggiungiPartecipante(this.getRegistroUtenti().getUtenteByEmail(emailPartecipante), appuntamento);
