+++ src/main/java/it/univaq/esc/repository/NotificaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.univaq.esc.model.notifiche.Notifica;
 
-public interface NotificaRepository extends JpaRepository<Notifica, Long>{
+public interface NotificaRepository extends JpaRepository<Notifica, Integer>{
 
 }
