+++ src/main/java/it/univaq/esc/repository/UtentePolisportivaAbstractRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import it.univaq.esc.model.utenti.UtentePolisportivaAbstract;
 
-public interface UtentePolisportivaAbstractRepository extends JpaRepository<UtentePolisportivaAbstract, Long> {
+public interface UtentePolisportivaAbstractRepository extends JpaRepository<UtentePolisportivaAbstract, Integer> {
     
 }
